requirements.txt
flask
gunicorn
google-genai
python-dotenv
Pillow
firebase-admin
google-cloud-storage
google-cloud-firestore
Werkzeug